requirements.txt
aiosmtplib==3.0.2
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
databases==0.8.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
fastapi-mail==1.5.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.25.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.6
liboqs-python @ file:///C:/QMsg/liboqs-python
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.19.0
pydantic==2.11.6
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.24.0
